float로 통일 일단은 더

+++ Sumdays/app/src/main/java/com/example/sumdays/data/dao/DailyEntryDao.kt
@@ -49,7 +49,7 @@ interface DailyEntryDao {
         diary: String? = null,
         keywords: String? = null,
         aiComment: String? = null,
-        emotionScore: Double? = null,
+        emotionScore: Float? = null,
         emotionIcon: String? = null,
         themeIcon: String? = null
     )
@@ -61,7 +61,7 @@ interface DailyEntryDao {
         diary: String? = null,
         keywords: String? = null,
         aiComment: String? = null,
-        emotionScore: Double? = null,
+        emotionScore: Float? = null,
         emotionIcon: String? = null,
         themeIcon: String? = null
     ) {
